Rework the milter interface and tweak the server to fit.
The milter interface was broken as far as protocol negotiation
went;  the protocol flags is supposed to be me sending all of
the phases I'm willing to send, and then that's pruned away
by the milter before they send back the phases they're willing
to accept (if they don't match, the milter connection fails
with a _completely meaningless_ log message; basically the
only "negotiation" is that I have to be willing to not do
any of the phases, and then send whatever ones the milter
is willing to accept.   Stupid design, but it's the standard.)

As a aide effect of this, I had to completely rework the
smtp server interface to account for milters failing out
at different places in the message.   I now have the milter
library failing all of the milters as soon as the first one
falls over dead, and then I can either reject the message
at that point (spam=bounce), defer the error handling until
the DATA phase (spam=folder), or just accept the tainted
mail (spam=accept) as is with the addition of a x-spam:
advertising header.

The `letter` structure now has a `healthy` flag, which
is set to 1 initially and zeroed if any milter raises
the freak flag.   If `let->healthy` is zero, I don't
need to call the milter for any other part of the smtp
transaction, because I already know it won't be accepted.